- change the way TachyOS initialize application environment  . user programs (threads) access only the copy of api struct instead of accessing itself

+++ source/arch/ARM/cortex-m4/tch_port.c
@@ -40,7 +40,7 @@
 #define PE_DYNAMIC					((uint8_t) 3)
 
 
-#define NR_PAGE_ENTRY				8
+#define NR_PAGE_ENTRY				((uint8_t) 8)
